Repositorys, Services

+++ SteelTech/src/main/java/com/back/SteelTech/Entity/Produto.java
@@ -9,9 +9,9 @@ import java.util.List;
 public class Produto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
     private String nome;
-    private Double preco;
+    private float preco;
     private String categoria;
 
     @OneToMany(mappedBy = "produto", cascade = CascadeType.ALL, orphanRemoval = true)
